requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.3
dj-database-url==3.1.0
django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
dotenv==0.9.9
gunicorn==24.1.1
packaging==26.0
psycopg==3.3.2
pycparser==3.0
pyjwt==2.10.1
python-dotenv==1.2.1
ruff==0.14.14
sqlparse==0.5.5
typing-extensions==4.15.0
whitenoise==6.11.0
